# speed up feature tests