So, 1min 30 is way too long for tests to run.
We are back under a minute

# before
```
Top 10 slowest examples (74.13 seconds, 81.9% of total time):
  home page user pages through comments
    19.95 seconds ./spec/features/home_page_spec.rb:17
  home page user pages through comments
    14.38 seconds ./spec/features/movie_page_spec.rb:12
  pages user creates, edits, and deletes a page
    13.64 seconds ./spec/features/pages_spec.rb:4
  HomeController GET charts gathers the information needed in the charts
    10.3 seconds ./spec/controllers/home_controller_spec.rb:32
  home page user searches for tv show
    4.29 seconds ./spec/features/home_page_spec.rb:30
  settings page user changes the footer
    3.58 seconds ./spec/features/settings_page_spec.rb:55
  settings page user changes the banner area
    3.48 seconds ./spec/features/settings_page_spec.rb:41
  settings page user changes the site name
    1.65 seconds ./spec/features/settings_page_spec.rb:5
  settings page user changes the sub-url
    1.49 seconds ./spec/features/settings_page_spec.rb:69
  home page user searches for a movie
    1.38 seconds ./spec/features/home_page_spec.rb:39
```
```
Finished in 1 minute 30.52 seconds (files took 1.84 seconds to load)
```
# after
```
Top 10 slowest examples (39.79 seconds, 67.9% of total time):
  pages user creates, edits, and deletes a page
    9.37 seconds ./spec/features/pages_spec.rb:4
  HomeController GET charts gathers the information needed in the charts
    6.64 seconds ./spec/controllers/home_controller_spec.rb:32
  home page user pages through comments
    6.45 seconds ./spec/features/home_page_spec.rb:17
  home page user pages through comments
    4.44 seconds ./spec/features/movie_page_spec.rb:12
  settings page user changes the footer
    3.51 seconds ./spec/features/settings_page_spec.rb:55
  settings page user changes the banner area
    3.49 seconds ./spec/features/settings_page_spec.rb:41
  settings page user changes the sub-url
    1.56 seconds ./spec/features/settings_page_spec.rb:69
  settings page user changes the site name
    1.49 seconds ./spec/features/settings_page_spec.rb:5
  home page user searches for a movie
    1.47 seconds ./spec/features/home_page_spec.rb:39
  settings page user changes the about page
    1.38 seconds ./spec/features/settings_page_spec.rb:14
```

```
Finished in 58.6 seconds (files took 1.95 seconds to load)
```