@HParker 2015-12-10
Still only takes 8 seconds to run the non feature tests on CI. I also didn't know Circle isn't running the feature tests.